add some commands, unfinished qr module

+++ modules/extensions/x_messaging.py
@@ -53,15 +53,20 @@ class Messaging(commands.Cog):
             `{0}delete 987654321987654321`, or
             reply with `{0}delete` to the message target
         """
-        async with ctx.typing():
-            if target:
+        # TODO: Delete these three lines for "delete" command
+        not_implemented = True
+        if not_implemented:
+            raise NotImplementedError("delete command is not implemented")
+        
+        if target:
+            async with ctx.typing():
                 # location = ctx.message.
                 if isinstance(target, str) and target.startswith("https://"):
                     target = target.split("/")[-1]
-
+    
                 # message = commands.MessageConverter.convert(ctx=ctx, argument=target)
                 # message.delete()
-                await ctx.message.delete()
+            await ctx.message.delete()
 
     # @commands.command()
     # async def up(self, ctx):
